+++ plugins/inventory/ntnx_prism_vm_inventory.py
@@ -180,7 +180,7 @@ import tempfile  # noqa: E402
 from ansible.errors import AnsibleError  # noqa: E402
 from ansible.plugins.inventory import BaseInventoryPlugin, Constructable  # noqa: E402
 
-from ..module_utils.utils import get_hostname  # noqa: E402
+from .inventory_utils import get_hostname  # noqa: E402
 from ..module_utils.v3.prism import vms  # noqa: E402
